Removed redundant test for ETCD in presubmit tests (#1861)
* Removed redundant test for ETCD in presubmit tests

Setting WITH_ETCD causes ETCD_DIR to be set, which is used in the integration tests. This env variable is not read in the presubmit tests though, which collapses the row to be equivalent to one of the existing flagsets. This will save Travis from running the presubmit tests with race detection twice per run, which is ~7m30s of execution time.

* Refactored the env section to remove more redundant artefacts.

The module global variables are the defaults in go v1.13 which Travis is already building with. No need to keep this around.