Factored test coverage generation into a Travis job
The integration tests weren't actually generating a coverage report anyway. The presubmit checks were, but the new job is equivalent to that. The two entries have been removed from the env configuration because they are now equivalent to the line under each of them, modulo the race flag. If we're going to run the tests with the race flag anyway, then there is no advantage to running the tests without it too; it's just a waste of compute time.

This PR tidies up the after_success which was running after all of our 16 tests, but only doing something for one of them. That's an unneccessary load on the codecov website, which I have observed timeouts in the travis logs for. Let's be good citizens.