Factored test coverage generation into a Travis job (#1862)
The integration tests weren't actually generating a coverage report anyway. The presubmit checks were, but the new job is equivalent to that. The two entries have been removed from the env configuration because they are now equivalent to the line under each of them, modulo the race flag. If we're going to run the tests with the race flag anyway, then there is no advantage to running the tests without it too; it's just a waste of compute time.

This PR tidies up the after_success which was running after all of our 16 tests, but only doing something for one of them. That's an unneccessary load on the codecov website, which I have observed timeouts in the travis logs for. Let's be good citizens.

The addition of the first env line not having coverage and race detection speeds up the execution time of Travis; this PR reduces the amount of load on Travis by ~1 min dev-perceived per build by shrinking the long pole, and ~15mins total compute time per build (1hr30m -> 1hr15m).