Only install softhsm and etcd in the test targets that need them
Only the integration tests for PKCS11 need softhsm installed, so skip
doing this for all of the other Travis test targets. Fixes #1851.

Only install etcd tools if we are running etcd in the tests. Installing
etcd tools takes >35s on some runs. This will save useless work for the
majority of targets.