Disable go build caching in Travis
TBD if this makes it faster or slower. It was taking ~20s after each build target to update the cache, and as all the builds for a revision happen in parallel and with different build tags, there are a lot of good reasons this costs more to move bytes than it saves us in compile time.