HStar3: A faster non-recursive HStar2 (#1874)
This change introduces the new HStar3 algorithm, a future replacement for HStar2. It takes advantage of the faster NodeID2 type operations like Prefix and Sibling, and doesn't use the heavy big.Int calculations.

Benchmarks compare with a 10-15% speed-up on the same dataset:

BenchmarkHStar2Root-12    	       2	 517407218 ns/op
BenchmarkHStar3Root-12    	       3	 443897809 ns/op

HStar3 code also claims to be more readable, as it is not recursive.

It is also more generic. For example, it doesn't depend on MapHasher interface and has no "empty hash" notion. The idea is that the caller is the one who provides empty hashes. It allows more flexibility: for example, the caller can error if some hash is not present (e.g. they used the GetReadSet method to prefetch all hashes, but HStar3 tries to read something different), or return an empty hash (e.g. if they did not use the prefetching option).

The lack of empty hash notion allows (in principle) using HStar3 even with logs. If HashChildrenFn returns nil unless both children are non-nil, and otherwise returns a "log" HashChildren then we can basically get similar functionality as compact.Range provides.