Fixed flakiness introduced in #1872
Verifying root correctness at old revisions isn't guaranteed to work because we only keep a finite cache of committed state. Trying to check SMRs for revisions > 10 in the past would fail. The previous code only checked this immediately on write, where #1872 performed the check on both a) examining the latest SMR; and b) examining historic SMRs. This change only keeps a) and drops the check in b), because the old state may have been purged. This gives less coverage, but is closer to what we had before. As work progresses on this hammer, I intend to increase coverage over historic roots. The priority for now is removing the flake though.

This caused flakiness because of the randomness of the hammer and the number of iterations. This change also increases the number of operations performed in Travis, which would turn this from a flakey failure to a persistent failure, if this PR didn't also include the medicine.