Refactor invalid ops into separate class (#1888)
* Refactor: moved invalid read ops into separate class
* Renamed methods on invalidReadOps to remove redundant naming
* Reduced duplication by introducing layers of indirection

This finishes off what was hinted at by the work started in #1886. This removes logic from the hammer class, and removes duplication of switching.

The next stage will be to break out the write functionality so that it is only called by the main thread, and the read methods will only be used by the checkers threads. This heads towards the design in #1880.