Refactor: removed PRNG from hammerState
This PR introduces a mapWorker class which has its own PRNG and will be scoped to a single threaded worker. I took the opportunity to move a number of methods from hammerState to mapWorker too. The goal is eventually that each worker would only have the local state it needs, plus the explicit thread-safe objects for sharing information - effectively an in-memory gossip network.