Added extra checks for bad states (#1899)
This makes the code do what the comment implied was going to happen. Any missing revisions would be bad news given how these previous contents are used for checking map correctness.

Also took the opportunity to simplify a little code that impaired readability.

None of this is triggered as the hammer is currently set up, but you may be surprised to learn that making the hammer properly concurrent will trigger some of these.