Refactor: extracted readWorker and writeWorker (#1893)
Reads and writes are logically distinct entities in the Map ecosystem. This refactor makes that explicit.

The next step is to make the read worker perform random read operations instead of just getLeavesRev. The next next step will be to make the write thread only perform writes.

This shapes things up in line with #1880.