Refactor: read worker uses MapBias, but maintains same behaviour (#1901)
It is clear from this point how the read worker will be changed to perform other read-only operations. This is a refactor and does not change behaviour except in one regard: the counter metrics for total operations performed are now incremented by the checker threads too.

Surely this is the last refactor before #1880 can be reached? :-)

In seriousness, there is a change that needs to happen to the data structures which maintain previous contents of the map and the SMRs to make them thread-safe. This will happen in a coming PR.