The Map Client now returns the SMR when fetching leaves
This allows clients to be sure which version of the map they are seeing. This is known in the case of GetAndVerifyMapLeavesByRevision, but for GetAndVerifyMapLeaves the caller was completely blind to this. It was causing problems in the Hammer, and it seems likely other clients will want this. If they don't, it can be easily ignored.

This is a breaking API change, but the Map is experimental and the CHANGELOG has been updated.