Hammer supports writer getting ahead of published SMR (#1907)
This will be required when WriteLeaves publishes SMRs asynchronously to writing the leaf values. This change makes VersionedMapContents updated with a revision X only when an SMR for a revision >= X is found. This allows the readers to keep using VersionedMapContents as they did before, without risking the writer getting so far ahead of the latest SMR that we push out all 10 of the revisions stored in memory that could be used to validate what the readers are getting from the map.