WriteLeaves: Always require the expected revision (#1912)
This change removes the feature of Set/WriteLeaves API calls, which allowed to
commit leaves to a new "latest" revision rather than a fixed one. This feature
is dangerous because it can cause races etc, and it was not used anyway.
Additionally, the revision == 0 meaning was overloaded (depending on the RPC
being called): it could mean the "empty tree" revision, or "the latest"
revision. Now it has only one meaning.