Read and write batch sizes are configured separately (#1923)
There is no reason in the real world that the number of leaves fetched by clients in a batch would have any relationship to the number of leaves written at a time. The single parameter was an artefact of the way the hammer was written. Allowing these to be configured separately better approximates the real world and allows the hammer to hit the map more flexibly.

Making either of the read or write flags be above 100 triggers the issue in #1845, though it does so differently for each of the batch sizes.