Replace legacy sparse Merkle tree writer (#1902)
This change removes SparseMerkleTreeWriter, and replaces it with a faster and
clearer smt.Writer.

The key design decision is to keep sharding/concurrency of smt.Writer on the
highest possible level (the application). It is unlikely that we would want to
move it to storage, as we might want to extend map_tree_updater.go to a
distributed setup, which is easier to do on the application level. Another
advantage is that map_tree_updater.go is not exported, which means we can change
this logic more easily.