Add supported deployment with Terraform
Create a root directory with a default supported deployment (based on
previous resources from examples/deployment). The intent is to have
this directory for all the supported deployment types and keep examples/
for deployments that are considered community-supported.

The first supported deployment is Trillian on Google Cloud using Cloud
Spanner as a backend.