Disabled python support in Bazel (#1964)
This was broken by an upstream change. We can only use master from repo because they don't have any tagged releases that are recent enough to build on a recognisable version of Bazel. Python support was changed in a recent commit upstream that breaks things for our usage. This is almost certainly fixable with enough time investigating, but AFAIK we have zero clients integrating by Bazel, and of this empty set, we know of zero Python users. Noted it in the CHANGELOG, and if anyone shouts we can prioritize.