Do not refund tokens for Tree- and User- scoped buckets. (#2025)
This prevents any user or log from being allowed infinite QPS if a reproducible failure is encountered. Failures can be more expensive than happy requests, and thus rate limiting them is fair game. The comments in the inteceptor strongly indicate that the refunding was only intended to be applied for global writes.

Fixes #2022.