Run `go generate ./...` on clean tip of tree
Update generated pb files from tip of tree. The changes here are a
result of changes in the dependency chain and tooling. No changes in
code are included in this PR.

Note that a manual update to grpc was done via `go get -u
google.golang.org/grpc` as per suggestion by
https://github.com/grpc/grpc-go/issues/3347#issuecomment-580922023.