smt: Introduce type for sorted list of nodes (#2129)
This change introduces NodesRow type which encapsulates a list of nodes with
unique and sorted IDs. This type is used for storing nodes in Tile type, and for
passing in updates to its Merge method. More usages will be added in follow-up
PRs.

So far NodesRow is backwards compatible with the previously used plain slice
[]Node.