merkle: Use compact range helpers for proofs (#2142)
This change removes the old code for generating proofs, and replaces it with a
much shorter version based on compact range helpers.

This code is 100% tested. To have >100% confidence, while working on this change,
I added an intermediary test which iterates over all possible parameters for inclusion
and consistency proof functions, and checks that the refactored code produces
results identical to the old functions.