Experimental beam library for verifiable maps (#2248)
This is a Beam library which has 2 modes of operation:
1) constructs a verifiable map from a set of key/value entries
2) updates a previously generated verifiable map with a delta of key/value entries

The tests pass so I feel pretty confident that this is useful in its current state, even though no demo code is provided. I have a demo locally which I've ran using the Python portable runner. I'll clean this up and submit in a follow-up PR.