Change demo to write out the tiles instead of a summary
Moved the demo script into a build directory and added verification demo. This verification shows what information is in the Tiles and is a closer approximation to real usage than the summarization. Still TODO is to generate/verify the inclusion proof, as the root hash is not checked in the demo yet.