Demo that creates a map from generated key/values (#2256)
This generates one file per tile for simplicity. This can only work while run on a single machine, although this does not confine it to be single-process. For this to run on a proper multi-machine cluster it would require a more suitable output sink.

There is a verifier provided which confirms inclusion of values by verifying the Merkle structure.

For those not afraid of the internals of Trillian, one cool aspect of this demo is that the Map is built using HStar3, but the verification is performed using HStar2.