log: Clean-up operation loop (#2294)
* Clean up getActiveLogIDs

* Remove unnecessary redundant glog

A few lines above there is glog printing the number of processed
entries, and processing time. There is not much sense in aggregating the
number of processed entries across log IDs for glog purposes, this info
is better represented in already existing metrics.

* Simplify reading from channel

* Simplify parallel execution of log operations

* Transpose the operations execution loop

This not only makes the code simpler, but prepares it for inverting
control in follow-up commits. Previously, workers invoked per-logID
functions, but after the refactoring each log will have a dedicated
worker.

* Factor out per-log execution

* Create mastership tracker in constructor

* Add period in comments

* Avoid "break label" by factoring out a method