client: Fix flaky WaitForInclusion test (#2327)
Sometimes the test fails with this error:

WaitForInclusion before sequencing: rpc error: code = Unknown desc = context deadline exceeded, want: not-nil
in the place where it expects status.DeadlineExceeded.

This is because not all Trillian gRPC methods return errors wrapped into
status. This change adds a workaround to unflake the test.