client: Fix flaky WaitForInclusion test
Sometimes the test fails with this error:

```
WaitForInclusion before sequencing: rpc error: code = Unknown desc =
sql: transaction has already been committed or rolled back, want:
not-nil
```

This could be because MySQL storage layer doesn't behave consistently
under a canceled Context.

For the purposes of the test it is sufficient to check that *some* error
is returned rather than a concrete cancelation error.