Upgraded to a version of etcd compatible with newer grpc
Trillian has been stuck on grpc < 1.30 - #2195. This uses the v3 version of etcd which isn't using the experimental (and now deleted) grpc APIs

I've deleted the prometheus etcdiscover tool in this commit. This can be re-added if we believe it's used, and we can find a way to test it. The upgrade path for this looks non-trivial.