Fix errors about missing go_package in proto files
Warnings are otherwise being printed saying that these proto files will not be supported in a future version of the compiler.

This requires more magic options being passed to protoc via gen.go files, but only along the same theme that we have in the other gen.go files in the codebase (without this the pb.go files are put in a hilariously nested directory starting `github.com/` rooted at the `gen.go` file, which seems like something that nobody would want, ever).