Wordsmithing/refactoring to make docs more accessible (#2494)
Improved the flow a little. Before introducing a Claimant Model using the template, I've included free text before it which explains in relatively plain English how the system functions. In the core model, this was mostly a case of moving the paragraph from after the template up to before it.