Remove WriteRevision from interfaces (#2534)
Existing storage implementations handle revisions internally. Revisions are removed from
the interfaces so that it becomes an implementation detail of each storage. This allows
more efficient implementations that don't require revisions.

This also deprecates the Revision field in the LogRootV1 structure. This field questionably
shouldn't have existed in the first place (layering violation), because revision was a storage
layer concept, and wasn't visible through the external API, whereas LogRootV1 is part of the
public API.