SubtreeCache: Batch tile reads when writing nodes (#2589)
The SubtreeCache uses two types of "read" callbacks: GetSubtreeFunc for
single-tile reads, and GetSubtreesFunc for batched reads. Besides being
complex, this also has performance issues: the single-node SetNodeHash
method fetches only one tile using GetSubtreeFunc, so a typical
SetMerkleNodes call results in a series of individual reads.

The only use-case for node writes in Trillian (the sequencer) does exactly one
SetMerkleNodes call per transaction with the entire write set, so there is no
need to sub-divide it into single-node writes in SubtreeCache.

This change replaces the SetNodeHash function with a batched SetNodes so
that the corresponding reads are batched too, and the single-tile callback is no
longer needed and removed.