Switch to regular maps in SubtreeCache (#2594)
The application-level logic calls GetMerkleNodes and SetMerkleNodes at
most once, and never does it concurrently. Correspondingly, the GetNodes
and SetNodes methods of SubtreeCache don't need to be thread-safe.

The thread safety of SubtreeCache was introduced for the Map needs, but
the Map mode has been deleted for now.

This change simplifies SubtreeCache so that it doesn't expect parallel calls.