Cloudbuild: Lock the golang tooling to a minor version (#2608)
* Lock the golang tooling to a minor version

Larger deltas in go tooling and expectations occur on minor version changes, which can cause problems for presubmits. Specifically, 1.16 to 1.17 has a change to how go:build lines are preferred, which is introducing diffs into PRs that are orthogonal. Locking this down means we now have an explicit decision about when we jump and address these changes.

* Updated documentation