Rewrite in-memory Merkle tree (#2700)
The code for internal/merkle/inmemory is old and needs a refresh. This change rewrites
the in-memory Merkle tree concisely, and updates all the call sites.

The previous implementation quoted C++ code, used a lot of non-Go constructs, and did
not have the benefit of reusing other Merkle primitives. All of these are good reasons for a
complete rewrite.