Remove cross-checking in integration test
Cross-checking between compact ranges and inmemory.Tree must be done in
unit tests for compact ranges or inmemory.Tree. Integration test seems
to be a wrong place to, essentially, inline this unit-test. At the level
of the integration test it would make more sense to cross-check with one
of the primitives.