Remake dumplib into integration test (#2710)
dumplib was a tool which created an in-memory log storage of a particular size,
and printed its contents with various flags. The useful bit of this tool was a "golden"
unit test which checks regressions on the SubtreeProto format in storage.

This change removes the tool, refactors the test, and places it in integration/format.