Rewrite in-memory Merkle tree tests (#2707)
This change completes the TODO of rewriting the unit tests for the in-memory Tree type.
The new tests are mostly a rework of the previous tests.

The methodology of testing remains the same:

- There is a reference implementation of root hash and proofs computation, translated from
definitions in RFC 6962. It is tested using "golden" tests and sanity checks.
- The Tree type is tested exhaustively, to make sure it produces the same result as the
reference implementation.

The goal is to move the Tree type and tests to https://github.com/transparency-dev/merkle
testonly package, as it can be widely reused for tests.