Removed support for Bazel (#2743)
To our knowledge nobody was using this and python support was already removed in #1964. If the removal of this affects anybody and they let us know, then I'm open to bringing this back. Without knowing it's used, this is something that breaks from time to time and demands investigation to keep CI working and thus hard to justify the return on investment.

Fixes #2690