Updated linter to latest version and fixed errors (#2748)
The spanner storage had the following error:
```
storage/cloudspanner/storage_provider.go:84:25: SA1019: r.SessionPoolConfig.MaxBurst is deprecated: MaxBurst exists for historical compatibility and should not be used. MaxBurst was used to limit the number of sessions that the session pool could create within a time frame. This was an early safety valve to prevent a client from overwhelming the backend if a large number of sessions was suddenly needed. The session pool would then pause the creation of sessions for a while. Such a pause is no longer needed and the implementation has been removed from the pool. (staticcheck)
	setUint64IfNotDefault(&r.SessionPoolConfig.MaxBurst, *csSessionMaxBurst)
```