CockroachDB: Optimize quota manager by using follower reads
This takes into use CockroachDB's follower reads feature [1] which
allows for the expensive `SELECT` query to be executed on a local node
instead of the leader. This reduces latency for multi-region
deployments; though it slows down the tests as we now allow for some
time for the database writes to converge.

[1] https://www.cockroachlabs.com/docs/stable/follower-reads.html#when-to-use-exact-staleness-reads

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>