Prototype tooling for Claimant Models (#2974)
The Claimant Model has proved itself as a useful theoretical tool for
understanding and communicating about the participants in a log
ecosystem. This tooling aims to make modeling & design be:
a) more useful
b) more interactive
c) less intimidating

This is still very much experimental, but I'd like other people to get
involved with the experiment and hack on this tooling to advance the
above goals.