mysql: check for error when getting subtrees
treeTX.getSubtrees iterates over a set of rows from `database/sql`, but
it was failing to check the error status after exiting the for loop that
iterates over them.

On our log deployment, this is causing mistaken error messages of
`preload did not get all tiles` when the real error should be propagated
from mysql (in our case, max_statement_time exceeded).