Add instructions for using docker to regen derived files
This is much cleaner than the previous approach of requiring protoc to be installed globally on the machine, which inevitably causes tensions with other projects that are developed on the same machine. We may want to consider building and publishing the Docker image in a repository so that people can just pull it and use it unless they are upgrading any of the tool versions (e.g. protoc).