Updated batchmap docs
I've just set this up on a new machine and confirmed it works. I've included the steps I took here because this is likely to work for other people, and it's all a bit mysterious without a script to follow.

Future work will involve updating this demo to actually explain what it is doing, because it's only just accessible to me, and I wrote it.