[Dependabot] Use multi-directory config for docker (#3671)
This removes a lot of duplication from the config. The problem I was trying to solve was making all of the docker deps be updated at the same time. This is what the 'groups' configuration block is supposed to do, but it doesn't seem to work in isolation. We'll see if this change in conjunction with that changes the update behaviour. If not, at least any further changes to the configs are done only in one place instead of in 7 almost identical blocks.