chore: relax go directive to permit 1.22.x
Setting this to 1.22.7 requires all consumers to be building with the
1.22.7 or newer release of Go 1.22 and to update their own go.mod
accordingly — this seems unnecessarily restrictive for a library module,
particularly as the code itself doesn't currently use any modern
language constructs and builds fine even with older Go versions.

Instead set the go directory to 1.22.0 and use the toolchain directive
to recommend the latest 1.22.x when building locally.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>