Ignore ErrTxDone when closing transactions (#3738)
Transactions can be closed in indirect ways, such as the context being cancelled. The old code would log an error when then occurred, but the new code checks for the transaction already being closed and treats this as a success. Fixes #3734 and fixes #2998. Note that this doesn't change the behaviour except for the log output spam.